Implemented chromosome classes for factory, city and delivery company entities.

+++ BusinessLogic/IChromosome.cs
@@ -6,6 +6,6 @@ namespace BusinessLogic
 {
     public interface IChromosome
     {
-        public float CalculateRank();
+        public double CalculateRank();
     }
 }
